sec commit

+++ app/src/main/java/com/example/mmvvmnew/ui/models/WeatherProperties.kt
@@ -4,11 +4,11 @@ import androidx.room.Entity
 
 data class WeatherProperties(
     val grnd_level: Double,
-    val humidity: Int,
+    val humidity: Double,
     val pressure: Double,
     val sea_level: Double,
     val temp: Double,
-    val temp_kf: Int,
+    val temp_kf: Double,
     val temp_max: Double,
     val temp_min: Double
 )
